+++ src/main/java/ua/kpi/diploma/controltestinghub/controller/DataSetController.java
@@ -90,7 +90,7 @@ public class DataSetController {
      */
     @PostMapping("/create-data-set")
     public void createDataSet(@RequestBody DataSetDto dataSetValues) {
-        long id = dataSetService.createDataSet(dataSetValues.getDataSetName());
+        Integer id = dataSetService.createDataSet(dataSetValues.getDataSetName());
         log.info("created data set id: {}, values: {}", id, dataSetValues.getDataEntryValues());
         dataEntryService.createDataEntry(id, dataSetValues.getDataEntryValues());
     }
@@ -111,7 +111,7 @@ public class DataSetController {
     }
 
     @GetMapping("/data-set/{id}/entries")
-    public List<DataEntry> getDatasetEntries(@PathVariable("id") Long id){
+    public List<DataEntry> getDatasetEntries(@PathVariable("id") Integer id){
         return dataEntryService.getAllByDataSetId(id);
     }
 }
