+++ src/main/java/ua/kpi/diploma/controltestinghub/service/ReportService.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface ReportService {
     ResponseEntity<?> sendReportToUser(List<ActionExecutionDto> actionExecutionList, List<SubscribedUserTestCaseDto> subscribedUsers);
-    List<SubscribedUserTestCaseDto> getSubscribedUsers(Long testCaseExecutionId);
+    List<SubscribedUserTestCaseDto> getSubscribedUsers(Integer testCaseExecutionId);
 }
