+++ src/main/java/ua/kpi/diploma/controltestinghub/service/VariableService.java
@@ -3,5 +3,5 @@ package ua.kpi.diploma.controltestinghub.service;
 import java.util.List;
 
 public interface VariableService {
-    void createVariables(Long actionId, List<String> names);
+    void createVariables(Integer actionId, List<String> names);
 }
